Docs: projects 등록 - 수정 트랜잭션 커밋 위치, 예외처리, 시간대 수정

+++ API/projects/controllers/project.controller.js
@@ -65,8 +65,6 @@ exports.project = async (req, res) => {
   const endMsec = Date.parse(end);
   if (startMsec >= endMsec) return res.status(400).json({ errorMessage: "프로젝트 기간 형식이 잘못되었습니다." });
 
-  const validSchedules = schedule.map((time) => moment(time, 'YYYY년MM월DD일 HH:mm').format("YYYY-MM-DD HH:mm:ss"));
-  
   const available = true;
   const createdAt = moment().format("YYYY-MM-DD HH:mm:ss");
   const email = userId;
@@ -75,11 +73,11 @@ exports.project = async (req, res) => {
 
   try {
     await Project.create({ title, details, subscript, role, start, end, email, id, nickname, createdAt }).then((result) => {
-      validSchedules.forEach((time) => Application.create({ projectId: result.projectId, schedule: time, available }));
+      schedule.forEach((time) => Application.create({ projectId: result.projectId, schedule: time, available }));
 
       skills.forEach((skill) => ProjectSkill.create({ projectId: result.projectId, skill }));
 
-      if (photos) {
+      if (photos && photos.length) {
         photos.forEach((photo) => ProjectPhoto.create({ projectId: result.projectId, photo }));
       };
     });
@@ -279,7 +277,7 @@ exports.projectUpdate = async (req, res) => {
     });
     
     const newAvailableApps = applications.filter((app) => app.available === 1 || app.available === true );
-    const validSchedules = newAvailableApps.map((app) => moment(app.schedule, 'YYYY년MM월DD일 HH:mm').format("YYYY-MM-DD HH:mm:ss"));
+    const validSchedules = newAvailableApps.map((app) => moment(app.schedule, 'YYYY-MM-DD HH:mm').format("YYYY-MM-DD HH:mm:ss"));
     const existUnavailableSchedules = existUnavailableApps.map((app) => app.schedule);
     const alreadyExistApps = validSchedules.filter((time) => existUnavailableSchedules.includes(time));
     
@@ -313,10 +311,8 @@ exports.projectUpdate = async (req, res) => {
       await ProjectSkill.create({ projectId, skill: skills[i] }, { transaction: t });
     };
 
-    await t.commit();
-
     // 예외처리 문제로 트랜잭션 밖으로 빼 줍니다.
-    if (photos || photos.length || !photos[0] === null) {
+    if (photos && photos.length && !photos[0] === null) {
       await ProjectPhoto.destroy({ where: { projectId } });
       for (let i = 0; i < photos.length; i++) {
         await ProjectPhoto.create({ projectId, photo: photos[i] });
@@ -328,6 +324,8 @@ exports.projectUpdate = async (req, res) => {
       if (response == 1) console.log("1 Redis key deleted");
       if (response == 2) console.log("2 Redis key deleted");
     });
+
+    await t.commit();
 
     return res.status(200).json({
       message: "프로젝트 게시글을 수정했습니다.",
